Tests/LibWeb/Text/input/HTML/link-element-username-password-percent-encoded.html

<a id="username-and-password" href="http://user%20name:pa%40ss%3Aword@www.ladybird.org"></a>
<script src="../include.js"></script>
<script>
    test(() => {
        const anchorElement = document.getElementById('username-and-password');
        println(anchorElement.username);
        println(anchorElement.password);
    })
</script>
