Tests/LibWeb/Text/expected/HTML/link-element-username-password-percent-encoded.txt

user%20name
pa%40ss%3Aword
